// Frontend/src/app/services/models/registration.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Event } from '../models/event';
import { User } from '../models/user';
export interface Registration {
  event?: Event;
  idRegist?: number;
  reservationDate?: string;
  status?: 'PENDING' | 'CONFIRMED' | 'CANCELED';
  user?: User;
}